requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.3.2
openai==1.75.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pypdfium2==4.30.1
python-multipart==0.0.20
python-pptx==1.0.2
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
XlsxWriter==3.2.3
